+++ CHANGELOG.md
@@ -69,7 +69,7 @@
 
 ## 1.8.2-RELEASE (2021-10-23)
 
-1. Fix poap attendee duration time
+1. Fix poap attendee duration time 
 2. Fix failed to send poaps message
 
 ## 1.8.1-RELEASE (2021-10-21)